Added Correios API to the project

+++ dm106CarlosDrury/Controllers/OrdersController.cs
@@ -11,6 +11,7 @@ using System.Web.Http.Description;
 using dm106CarlosDrury.Models;
 using dm106CarlosDrury.CRMClient;
 using dm106CarlosDrury.br.com.correios.ws;
+using System.Globalization;
 
 namespace dm106CarlosDrury.Controllers
 {
@@ -23,25 +24,68 @@ namespace dm106CarlosDrury.Controllers
         [ResponseType(typeof(string))]
         [HttpGet]
         [Route("frete")]
-        public IHttpActionResult CalculaFrete()
+        public IHttpActionResult CalculaFrete(int id)
         {
-            string frete;
 
-            CalcPrecoPrazoWS correios = new CalcPrecoPrazoWS();
+            Order order = db.Orders.Find(id);
 
-            cResultado resultado = correios.CalcPrecoPrazo("", "", "40010", "37540000", "37002970", "1", 1, 30, 30, 30, 30, "N", 100, "S");
-
-            if (resultado.Servicos[0].Erro.Equals("0"))
+            if (order == null)
             {
-                frete = "Valor do frete: " + resultado.Servicos[0].Valor + " - Prazo de entrega: " + resultado.Servicos[0].PrazoEntrega + " dia(s)";
-                return Ok(frete);
+                return NotFound();
             }
             else
             {
-                return BadRequest("Código do erro: " + resultado.Servicos[0].Erro + "-" + resultado.Servicos[0].MsgErro);
+                CRMRestClient crmClient = new CRMRestClient();
+                Customer customer = crmClient.GetCustomerByEmail(order.emailUser);
+                if (customer != null)
+                {
+                    // Initial Values
+                    decimal larg = 0;
+                    decimal comp = 0; 
+                    decimal height = 0;
+                    order.totalWeigth = 0;
+                    order.totalPrice = 0;
+
+                    // Calculate size for each orderItem
+                    foreach (OrderItem orderItem in order.OrderItems)
+                    {
+                        int qtd = orderItem.qtd;
+                        height = orderItem.Product.altura > height ? orderItem.Product.altura : height;
+                        larg = orderItem.Product.largura > larg ? orderItem.Product.largura : larg;
+                        comp += orderItem.Product.comprimento * qtd;
+                        order.totalWeigth += orderItem.Product.peso * qtd;
+                        order.totalPrice += orderItem.Product.preco * qtd;
+                    }
+
+                    // Correios API
+                    string frete;
+                    CalcPrecoPrazoWS correios = new CalcPrecoPrazoWS();
+                    cResultado resultado = correios.CalcPrecoPrazo("", "", "40010", "37540000", customer.zip, order.totalWeigth.ToString(), 1, comp, height, larg, 0, "N", 100, "S");
+                    if (resultado.Servicos[0].Erro.Equals("0"))
+                    {
+                        NumberFormatInfo numberFormat = new NumberFormatInfo();
+                        numberFormat.NumberDecimalSeparator = ",";
+                        order.shipmentPrice = Decimal.Parse(resultado.Servicos[0].Valor, numberFormat);
+                        order.deliveryDate = order.orderDate.AddDays(Int32.Parse(resultado.Servicos[0].PrazoEntrega));
+                        frete = "Valor do frete: " + order.shipmentPrice + " - Prazo de entrega: " + order.deliveryDate;
+                        db.SaveChanges();
+
+                        return Ok(frete);
+                    }
+                    else
+                    {
+                        return BadRequest("Código do erro: " + resultado.Servicos[0].Erro + "-" + resultado.Servicos[0].MsgErro);
+                    }
+                }
+                else
+                {
+                    return BadRequest("Falha ao consultar o CRM");
+                }
             }
+                    
         }
 
+           
         [ResponseType(typeof(string))]
         [HttpGet]
         [Route("cep")]
@@ -60,6 +104,38 @@ namespace dm106CarlosDrury.Controllers
             }
         }
 
+        // GET: api/Orders/5
+        [ResponseType(typeof(Order))]
+        [HttpGet]
+        [Route("close")]
+        public IHttpActionResult CloseOrder(int id)
+        {
+            Order order = db.Orders.Find(id);
+
+            if (order == null)
+                return NotFound();
+
+            if (order.emailUser == User.Identity.Name || User.IsInRole("ADMIN"))
+            {
+                // order found
+                if (order.shipmentPrice != 0)
+                {
+                    order.status = "fechado";
+                    db.SaveChanges();
+                    return Ok("Order Closed");
+                }
+                else
+                {
+                    return BadRequest("Cannot close order due to shipment price was not defined.");
+                }
+            }             
+            else 
+            {
+                return BadRequest("Authorization Denied! Only admin or the order owner allowed!");
+            }
+           
+                
+        }
 
         // GET: api/Orders
         public List<Order> GetOrders()
@@ -79,53 +155,78 @@ namespace dm106CarlosDrury.Controllers
         public IHttpActionResult GetOrder(int id)
         {
             Order order = db.Orders.Find(id);
-            if (order == null)
-            {
+            if (order != null && (order.emailUser == User.Identity.Name || User.IsInRole("ADMIN")))
+                return Ok(order);
+            else
                 return NotFound();
-            }
-
-            return Ok(order);
         }
 
         // PUT: api/Orders/5
         [ResponseType(typeof(void))]
         public IHttpActionResult PutOrder(int id, Order order)
         {
-            if (!ModelState.IsValid)
+            if(order != null && (order.emailUser == User.Identity.Name || User.IsInRole("ADMIN")))
             {
-                return BadRequest(ModelState);
-            }
-
-            if (id != order.Id)
-            {
-                return BadRequest();
-            }
-
-            db.Entry(order).State = EntityState.Modified;
-
-            try
-            {
-                db.SaveChanges();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!OrderExists(id))
+                if (!ModelState.IsValid)
                 {
-                    return NotFound();
+                    return BadRequest(ModelState);
                 }
-                else
-                {
-                    throw;
-                }
-            }
 
-            return StatusCode(HttpStatusCode.NoContent);
+                if (id != order.Id)
+                {
+                    return BadRequest();
+                }
+
+                db.Entry(order).State = EntityState.Modified;
+
+                try
+                {
+                    if (order != null && (order.emailUser == User.Identity.Name || User.IsInRole("ADMIN")))
+                    {
+                        db.SaveChanges();
+                    }
+                    else
+                    {
+                        return BadRequest();
+                    }
+
+                }
+                catch (DbUpdateConcurrencyException)
+                {
+                    if (!OrderExists(id))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
+                }
+                return StatusCode(HttpStatusCode.NoContent);
+            }
+            else
+            {
+                return BadRequest("Authorization Denied! Only admin or the order owner allowed!");
+            }
+                
+
+            
         }
 
         // POST: api/Orders
         [ResponseType(typeof(Order))]
         public IHttpActionResult PostOrder(Order order)
         {
+
+            // initialize order
+            order.status = "novo";
+            order.totalWeigth = 0;
+            order.shipmentPrice = 0;
+            order.totalPrice = 0;
+            order.orderDate = DateTime.Now;
+            order.emailUser = User.Identity.Name;
+
+
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
@@ -142,15 +243,23 @@ namespace dm106CarlosDrury.Controllers
         public IHttpActionResult DeleteOrder(int id)
         {
             Order order = db.Orders.Find(id);
+
             if (order == null)
-            {
                 return NotFound();
+
+            if (order.emailUser == User.Identity.Name || User.IsInRole("ADMIN"))
+            {
+                db.Orders.Remove(order);
+                db.SaveChanges();
+
+                return Ok(order);
+            }
+            else
+            {
+                return BadRequest("Authorization Denied! Only admin or the order owner allowed!");
             }
 
-            db.Orders.Remove(order);
-            db.SaveChanges();
 
-            return Ok(order);
         }
 
         protected override void Dispose(bool disposing)
